backend/services/custom_api.py
#!/usr/bin/env python3
"""
┌─────────────────────────────────────────────────────────────────────────┐
│                   CUSTOM API INTEGRATION PLACEHOLDER                    │
├─────────────────────────────────────────────────────────────────────────┤
│                                                                         │
│  This module is designed for easy integration with your custom         │
│  AI text detection API.                                                │
│                                                                         │
│  IMPLEMENTATION GUIDE:                                                  │
│  ────────────────────                                                   │
│                                                                         │
│  1. UPDATE CONFIGURATION (config/settings.py):                         │
│     - Set CUSTOM_API_KEY in .env                                       │
│     - Set CUSTOM_API_URL in .env                                       │
│                                                                         │
│  2. IMPLEMENT analyze_text() METHOD:                                    │
│     - Add your API authentication logic                                 │
│     - Build the API request with text content                           │
│     - Parse the API response                                            │
│     - Map response to standardized format                               │
│                                                                         │
│  3. RESPONSE FORMAT:                                                    │
│     {                                                                   │
│         'status': 'AUTHENTIC' or 'AI_GENERATED',                       │
│         'score': float (0.0 to 1.0),                                   │
│         'confidence': str (optional),                                   │
│         'metadata': dict (optional additional data)                     │
│     }                                                                   │
│                                                                         │
└─────────────────────────────────────────────────────────────────────────┘
"""

import logging
import requests
from typing import Dict, Any

logger = logging.getLogger(__name__)


class CustomTextAPIError(Exception):
    """Base exception for Custom API errors"""
    pass


class CustomTextAPIService:
    """
    Custom API Service for AI Text Detection
    
    TODO: Implement your custom API integration here
    """
    
    def __init__(
        self,
        api_key: str,
        api_url: str,
        timeout: int = 30,
        verify_ssl: bool = True
    ):
        """Initialize Custom API Service"""
        self.api_key = api_key
        self.api_url = api_url.rstrip('/')
        self.timeout = timeout
        self.verify_ssl = verify_ssl
        
        logger.info(f"CustomTextAPIService initialized with endpoint: {self.api_url}")
    
    def analyze_text(self, text: str) -> Dict[str, Any]:
        """
        Analyze text for AI-generated content
        
        ┌──────────────────────────────────────────────────────────────┐
        │ TODO: IMPLEMENT YOUR API INTEGRATION HERE                    │
        └──────────────────────────────────────────────────────────────┘
        
        Example Implementation:
        
        ```python
        # 1. Prepare request
        headers = {
            'Authorization': f'Bearer {self.api_key}',
            'Content-Type': 'application/json'
        }
        
        payload = {
            'text': text,
            'detailed': True
        }
        
        # 2. Send request
        response = requests.post(
            f'{self.api_url}/detect',
            headers=headers,
            json=payload,
            timeout=self.timeout
        )
        response.raise_for_status()
        
        # 3. Parse response
        data = response.json()
        
        # 4. Return standardized format
        return {
            'status': 'AI_GENERATED' if data['is_ai'] else 'AUTHENTIC',
            'score': float(data['confidence']),
            'confidence': self._calculate_confidence(data['confidence']),
            'metadata': data
        }
        ```
        """
        
        raise NotImplementedError(
            "\n\n"
            "╔════════════════════════════════════════════════════════════╗\n"
            "║  CUSTOM API NOT IMPLEMENTED                                ║\n"
            "╠════════════════════════════════════════════════════════════╣\n"
            "║                                                            ║\n"
            "║  Please implement the analyze_text() method in:           ║\n"
            "║  backend/services/custom_api.py                            ║\n"
            "║                                                            ║\n"
            "║  For now, the system is using MOCK mode.                  ║\n"
            "║                                                            ║\n"
            "╚════════════════════════════════════════════════════════════╝\n"
        )
    
    def _calculate_confidence(self, score: float) -> str:
        """Calculate confidence level from score"""
        if score >= 0.8:
            return "High"
        elif score >= 0.5:
            return "Medium"
        else:
            return "Low"
    
    def health_check(self) -> Dict[str, Any]:
        """Check if API is accessible"""
        return {
            'healthy': True,
            'message': 'Health check not implemented',
            'endpoint': self.api_url
        }
