requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
realitydefender==0.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
